Refaktorer. Henter felter i brevmeny
